Add files via upload
Accelerate Bayesian inference in function space using a neural operator (DeepONet). The neural operator surrogate model is trained to learn the function-space map between torque-vs time and the solution trajectory given some initial conditions of a nonlinear pendulum system. Trained on the angular position only, it can predict the trajectory based on arbitrary torque/time functions. The forward pass of DeepONet is much faster than the ODE/PDE/Optimizer solvers used to generate the data on which it trains, so inference within Turing or Gen can be much faster.